# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Um repositorio de autores e suas obras",
        "description": "Um repositorio para agrupar todos os autores e suas obras conhecidas da humanidade",
        "termsOfService": "https://minhaapi.com/terms",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/"
        }
    ],
    "paths": {
...
        "/Livro": {
            "get": {
                "tags": [
                    "Livro"
                ],
                "operationId": "8907fa605d04e363054cc6eda312d2de",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "titulo": {
                                                "description": "nome do livro",
                                                "format": "string",
                                                "default": "senhor dos aneis"
                                            },
                                            "genero": {
                                                "description": "genero do livro",
                                                "format": "string",
                                                "default": "fantasia"
                                            },
                                            "autor_id": {
                                                "description": "id do autor do livro",
                                                "format": "integer",
                                                "default": "1"
                                            },
                                            "numeroPaginas": {
                                                "description": "numero de paginas do livro",
                                                "format": "string",
                                                "default": "2000"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Livro"
                ],
                "summary": "Criar um Livro",
                "description": "Cria um Livro",
                "operationId": "fb54096a5a7748ae04b6a4d35d1d5225",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "genero": {
                                        "type": "string"
                                    },
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "numeroPaginas": {
                                        "type": "integer"
                                    },
                                    "autor_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "titulo": {
                                            "description": "nome do livro",
                                            "format": "string",
                                            "default": "senhor dos aneis"
                                        },
                                        "genero": {
                                            "description": "genero do livro",
                                            "format": "string",
                                            "default": "fantasia"
                                        },
                                        "autor_id": {
                                            "description": "id do autor do livro",
                                            "format": "integer",
                                            "default": "1"
                                        },
                                        "numeroPaginas": {
                                            "description": "numero de paginas do livro",
                                            "format": "string",
                                            "default": "2000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/Livro/{Livro}": {
            "get": {
                "tags": [
                    "Livro"
                ],
                "operationId": "5df9bf654b37d5003e4915faed85c96b",
                "parameters": [
                    {
                        "name": "Livro",
                        "in": "path",
                        "description": "Id do livro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "titulo": {
                                            "description": "nome do livro",
                                            "format": "string",
                                            "default": "senhor dos aneis"
                                        },
                                        "genero": {
                                            "description": "genero do livro",
                                            "format": "string",
                                            "default": "fantasia"
                                        },
                                        "autor_id": {
                                            "description": "id do autor do livro",
                                            "format": "integer",
                                            "default": "1"
                                        },
                                        "numeroPaginas": {
                                            "description": "numero de paginas do livro",
                                            "format": "string",
                                            "default": "2000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Livro"
                ],
                "summary": "Atualizar um Livro",
                "description": "Atualizar um Livro",
                "operationId": "9d2ee9f67449fa174123a34cdd66034a",
                "parameters": [
                    {
                        "name": "Livro",
                        "in": "path",
                        "description": "Id do livro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "genero": {
                                        "type": "string"
                                    },
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "numeroPaginas": {
                                        "type": "integer"
                                    },
                                    "autor_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "titulo": {
                                            "description": "nome do livro",
                                            "format": "string",
                                            "default": "senhor dos aneis"
                                        },
                                        "genero": {
                                            "description": "genero do livro",
                                            "format": "string",
                                            "default": "fantasia"
                                        },
                                        "autor_id": {
                                            "description": "id do autor do livro",
                                            "format": "integer",
                                            "default": "1"
                                        },
                                        "numeroPaginas": {
                                            "description": "numero de paginas do livro",
                                            "format": "string",
                                            "default": "2000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Livro"
                ],
                "summary": "Remove o Livro especifico",
                "description": "Remove o livro",
                "operationId": "92d3343ae6e305a3492f5a3c505a0ec5",
                "parameters": [
                    {
                        "name": "Livro",
                        "in": "path",
                        "description": "Id do livro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "string": {
                                            "description": " string indicando a mensagem",
                                            "format": "string",
                                            "default": " msg"
                                        },
                                        "": {
                                            "description": "mensagem falando que o Livro foi deletado",
                                            "format": "string",
                                            "default": "Livro deletado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Pegue o token indo na operação login e passando o email e senha",
                "name": "Token JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}